Move initialization to Initializer object
