+++ src/main/kotlin/frc/robot/Constants.kt
@@ -84,69 +84,4 @@ object Constants {
             { i -> GeometryUtil.flipFieldPose(chainLocations[i])}
         }
     }
-
-    fun initSwerve() {
-        var moduleIOs: Array<ModuleIO>
-
-        when (SwerveConstants.SWERVE_TYPE) {
-            SwerveConstants.SwerveType.SIM -> {
-                moduleIOs = Array<ModuleIO>(4) { ModuleIOSim() }
-                SwerveDrive.initialize(GyroIOSim(), SwerveConstants.OFFSETS, *moduleIOs)
-            }
-
-            SwerveConstants.SwerveType.WCP -> {
-                moduleIOs = Array<ModuleIO>(4) { i ->
-                    ModuleIOTalonFX(
-                        Ports.SwerveDriveWCP.DRIVE_IDS[i],
-                        Ports.SwerveDriveWCP.ANGLE_IDS[i],
-                        Ports.SwerveDriveWCP.ENCODER_IDS[i],
-                        SwerveConstants.DRIVE_MOTOR_CONFIGS
-                            ?: throw IllegalStateException("drive motor config is null"),
-                        SwerveConstants.ANGLE_MOTOR_CONFIGS
-                            ?: throw IllegalStateException("angle motor config is null"),
-                        SwerveConstants.ENCODER_CONFIGS ?: throw IllegalStateException("encoder config is null")
-                    )
-                }
-                SwerveDrive.initialize(GyroIOReal(), SwerveConstants.OFFSETS, *moduleIOs)
-            }
-
-            SwerveConstants.SwerveType.NEO -> {
-                moduleIOs = Array<ModuleIO>(4) { i ->
-                    ModuleIOSparkMax(
-                        Ports.SwerveDriveNEO.DRIVE_IDS[i],
-                        Ports.SwerveDriveNEO.ANGLE_IDS[i],
-                        Ports.SwerveDriveNEO.ENCODER_IDS[i],
-                        Ports.SwerveDriveNEO.DRIVE_INVERTED[i],
-                        Ports.SwerveDriveNEO.ANGLE_INVERTED[i]
-                    )
-                }
-                SwerveDrive.initialize(GyroIOReal(), SwerveConstants.OFFSETS, *moduleIOs)
-            }
-        }
-    }
-
-    fun initVision(){
-        var speakerRightCamera =
-            PhotonVisionIOReal(
-                PhotonCamera("OV2311_1"),
-                VisionConstants.SPEAKER_RIGHT_CAMERA_POSE
-            )
-        var speakerLeftCamera =
-            PhotonVisionIOReal(
-                PhotonCamera("OV2311_2"),
-                VisionConstants.SPEAKER_LEFT_CAMERA_POSE,
-            )
-        var intakeAprilTagCamera =
-            PhotonVisionIOReal(
-                PhotonCamera("OV2311_0"),
-                VisionConstants.INTAKE_APRILTAG_CAMERA_POSE,
-            )
-        var driverCamera =
-            PhotonVisionIOReal(
-                PhotonCamera("Driver_Camera"),
-                VisionConstants.DRIVER_CAMERA_POSE,
-            )
-
-        Vision.initialize(listOf(speakerRightCamera, speakerLeftCamera, intakeAprilTagCamera, driverCamera))
-    }
 }
